+++ ophelos_sdk/resources/__init__.py
@@ -4,6 +4,7 @@ Resource managers for Ophelos API endpoints.
 
 from .base import BaseResource
 from .communications import CommunicationsResource
+from .contact_details import ContactDetailsResource
 from .customers import CustomersResource
 from .debts import DebtsResource
 from .invoices import InvoicesResource
@@ -18,6 +19,7 @@ __all__ = [
     "BaseResource",
     "DebtsResource",
     "CustomersResource",
+    "ContactDetailsResource",
     "OrganisationsResource",
     "PaymentsResource",
     "InvoicesResource",